Move the download bits from TAP 4 to TAP 5.
Remove restrictions of how files are kept on repositories and clients.

Remove restriction on the file names of delegated targets roles in the
snapshot metadata. This came up in a meeting where it was thought that
the top-level targets role, and all delegated targets roles, should be
listed as if they were in the same directory, in order to prevent
delegated targets roles from accidentally overwriting top-level roles.

Since these restrictions were lifted, backwards compatibility looks
quite good.