Remove Test Battery section. Save minority of it elsewhere.
Move the notes on how the Tester will use the Wrapper functions
to the Wrapper Specification instead, integrating and rewording.

Move some basic text about what test cases are like to the
Tester Specification section.

Text on particular test cases is removed and will likely be
added to another document in the future.