Changes in Tap 4 specification, mechanism and steps
After working with the diagram, I went back to the text to try to
simplify and to add some specificity to the Specification section. One
principal change is since we are referring to “mappings” as the
individual list of repos, files etc, i’m using the verb “assign” to
describe how the mapping is done. I think I clarify the difference but
wanted to call out so you can check.