Remove Use Case 3, which is covered by 1 and 4 (now 3)
Fixes #88.

Use Cases 1 and 4 cover clients selecting which repository to trust for which
targets and trusting a combination of multiple repositories for a given target,
and so I think they already cover what Use Case 3 intends to demonstrate.

I think Use Case 3 just confuses things. If it's about client choice, that's
covered in 1 and 4, and if it's about repository maintainer choice, then that's
actually provided by TAP 3, counter to what Use Case 3 suggests.