Remove unneeded level of new format, 'roles' in 'delegations'
This seems to have been left in by mistake. The 'delegations'
element is now simply a list, each element of which is a
dictionary containing the data for a single delegation.

Instead, 'delegations' was a dictionary always containing a
single item called 'roles', which was the list described above.

Aside from being unnecessary, 'roles' was also a misleading
name for a list of delegations.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>