TAP 1: add TAP types - Standardization and Informational
Informational TAPs don't require an augmented reference implementation, or
any changes to the TUF specification, may be long-lived (with a
status of Active) and may receive updates. For example TAP 1 and TAP 11.

Standardization TAPs require a PoC to move from Draft to Accepted and will
not move to Final status until there's an implementation merged into
python-tuf and a corresponding change included in the specification.