TAP 15: improve examples story
Use the same number of bins, i.e. 2048, used for overhead calcs in
both example snippets for a coherent story.

Further improvements include in the  traditional hash bin
delegation example include:

- list the first two and the last bin, instead of the first 4.
This has the advantage that the bin index in the bin name, and the
path hash prefixes can be clearly identified as hex numbers, it
also shows better how many entries there are in total.
- spell out the values in `path_hash_prefixes`
Previously they were abbreviated with with an asterisk, which was
confusing as the TUF spec also has fields that support glob
patterns, but this is not one of them. Spelling them out also
stresses emphasis on the metadata overhead that can be saved with
succinct hash bin delegation.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>