TAP 15: Compare relative PEP 458 metadata overhead
Prior to this change we compared the overall metadata overhead of
traditional and succinct hash bin delegation (hbd), using different
numbers of bins optimized for each hashing scheme for a "fair"
comparison. However, using different numbers makes it harder to
understand the advantage of succinct hbd, i.e. no growth of the
delegating metadata vs. linear growth with number of bins.

This change focuses on the different sizes of the delegating
targets metadata for the same number of bins to underline the
advantage of succinct hbd.

Background info: The optimum number of bins for a given number of
target files varies between the two schemes, because in traditional
hbd the positive effect of many bins is undermined by the huge size
of the delegating metadata.

Other changes include:
- use a real world number of bins (PEP 458) instead of fictional
- compare metadata overhead percentages instead of absolute numbers
- perform keyid calcs with 64 instead of 32 bytes to match
  reference implementation
- remove mention of inaccurate (32-bit) and not so relevant keyid size

See
https://github.com/theupdateframework/taps/pull/155#discussion_r893501418
for initial discussion.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>
Co-authored-by: Joshua Lock <jlock@vmware.com>